feat(client): Add GoUpButton Component
